Rewrite as kubernetes operator/controller
This changes the approach of helm-state metrics from reading the current
cluster state (e.g. every helm release secret) on every scrape to
running a kubernetes controller that watches for changes to helm release
secrets and keeps state in prometheus registry.

This should greatly reduce the number of "full scans" of secrets from
the kubernetes API and therefore should stop the linear increase of API
response time with number of helm releases in a cluster.

Bug: T323706
Change-Id: I063d4530cf6b0259998597f3a6a9b33dda5f8ea7